patch changes to meet timing requirement

+++ src/Accelerators/MyAccel/Runtime/Conv2DKernel.h
@@ -8,11 +8,10 @@ extern "C" {
 #endif
 
 void conv2d_kernel(const float *x, const float *weight, const float *bias,
-    float *y, int64_t n_size, int64_t c_size, int64_t h_size,
-    int64_t input_w_size, int64_t m_size, int64_t kh_size, int64_t kw_size,
-    int64_t oh_size, int64_t ow_size, int64_t dilation_h, int64_t dilation_w,
-    int64_t c_per_group, int64_t group, int64_t pad_left, int64_t pad_top,
-    int64_t stride_h, int64_t stride_w, int64_t has_bias);
+    float *y, int n_size, int c_size, int h_size, int input_w_size,
+    int m_size, int kh_size, int kw_size, int oh_size, int ow_size,
+    int dilation_h, int dilation_w, int c_per_group, int group, int pad_left,
+    int pad_top, int stride_h, int stride_w, int has_bias);
 
 #ifdef __cplusplus
 }
